+++ srcs/operation_reverse_rotate.c
@@ -2,9 +2,9 @@
 
 void	rra(t_info *info)
 {
-	int	tmp_a;
-	int	tmp_compression_a;
-	int	i;
+	int		tmp_a;
+	int		tmp_compression_a;
+	size_t	i;
 
 	ft_putendl_fd("rra", 1);
 	i = 0;
@@ -27,9 +27,9 @@ void	rra(t_info *info)
 
 void	rrb(t_info *info)
 {
-	int	tmp_b;
-	int	tmp_compression_b;
-	int	i;
+	int		tmp_b;
+	int		tmp_compression_b;
+	size_t	i;
 
 	ft_putendl_fd("rrb", 1);
 	i = 0;
